pyproject.toml
[project]
name = "ai-mcp-pic-obj-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.12.0",
    "boto3>=1.42.24",
    "fastmcp>=2.14.2",
    "isort>=7.0.0",
    "langchain>=1.2.3",
    "langchain-aws>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.5",
    "langgraph-checkpoint>=3.0.1",
    "mcp[cli]>=1.25.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
]
